requirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
pandas
python-dotenv
pypdf
sentence-transformers
requests
google-generativeai
